# backend/brewery-api/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Product:
    description: Product defines the structure for an API product
    properties:
      description:
        description: the description for this poduct
        maxLength: 255
        type: string
        x-go-name: Description
      id:
        description: the id for this product
        format: int64
        type: integer
        x-go-name: ID
      name:
        description: the name for this poduct
        maxLength: 255
        type: string
        x-go-name: Name
      price:
        description: the price for the product
        format: float
        minimum: 0.01
        type: number
        x-go-name: Price
      sku:
        description: the SKU for the product
        pattern: '[a-z]+-[a-z]+-[0-9]+'
        type: string
        x-go-name: SKU
    required:
    - id
    - name
    - description
    - price
    - sku
    type: object
    x-go-package: brewery/api/data
info:
  description: Documentation for Product API
  title: of Product API
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns a list of products from the database
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
    post:
      description: Creates a new product
      operationId: createProduct
      responses:
        "201":
          description: The product was created successfully.
        "404":
          description: Product not found
      tags:
      - products
  /products/{id}:
    delete:
      description: Deletes product with the given id
      operationId: deleteProduct
      parameters:
      - description: The id of the product
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "204":
          description: The product was deleted successfully.
        "404":
          description: Product not found
      tags:
      - products
    get:
      description: Returns a single product from the database
      operationId: listProduct
      parameters:
      - description: The id of the product
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "200":
          $ref: '#/responses/productResponse'
        "404":
          description: Product not found
      tags:
      - products
  /products{id}:
    put:
      description: Updates product with the given id
      operationId: updateProduct
      parameters:
      - description: The id of the product
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "204":
          description: The product was updated successfully.
        "404":
          description: Product not found
      tags:
      - products
produces:
- application/json
responses:
  productResponse:
    description: a product returned in the response
    schema:
      $ref: '#/definitions/Product'
  productsResponse:
    description: A list of products returns in the response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"